Move HoverTooltip into its own class.
